LuaJIT: change function signature from ssize_t to int
Prior to LuaJIT 2.1.0-beta, LuaJIT had no support for ssize_t
declarations. Since we have a single declaration of a ssize_t member
it makes more sense to replace it with an int (which provides enough
precision for the data that member represents).

+++ src/lua_backend.cpp
@@ -69,7 +69,7 @@ extern "C" void *luaGetData(const char *element)
     return NULL;
 }
 
-extern "C" ssize_t luaGetElementSize(const char *element)
+extern "C" int luaGetElementSize(const char *element)
 {
     int index = index_of(element);
     if (index >= 0)
