+++ src/udf_template.lua
@@ -16,7 +16,7 @@ function init(filterpath)
         const char *luaGetType(const char *);
         const char *luaGetCast(const char *);
         const char *luaGetDims(const char *);
-        ssize_t     luaGetElementSize(const char *);
+        int         luaGetElementSize(const char *);
         // compound_declarations_placeholder
     ]]
